refactor: extract cli detection logic
